jarvis/voice_listener/requirements.txt
faster-whisper
sounddevice
soundfile
numpy
python-dotenv
